Replace Float by Double
git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@4370 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ app-ngl-data/app/services/description/sample/SampleService.java
@@ -120,7 +120,7 @@ public class SampleService {
 	public static List<PropertyDefinition> getPropertyDefinitionsADNGenomic() throws DAOException {
 		List<PropertyDefinition> propertyDefinitions = new ArrayList<PropertyDefinition>();
         propertyDefinitions.addAll(getPropertyDefinitionsADN());
-        propertyDefinitions.add(newPropertiesDefinition("%GC", "gcPercent", LevelService.getLevels(Level.CODE.Sample),Float.class, false, "single"));
+        propertyDefinitions.add(newPropertiesDefinition("%GC", "gcPercent", LevelService.getLevels(Level.CODE.Sample),Double.class, false, "single"));
 		return propertyDefinitions;
 	}
